Update nifti_to_mat_loop.m
This script will loop through a directory to find a specific folder name that contains either .nii.gz or .nii files then convert them to _roi.mat files. The only things to change are the named folder which is the folder containing the images (this should be the same for all subjects if running multiple) and the root directory.